+++ sopa/utils/__init__.py
@@ -1,0 +1,1 @@
+from .image import get_channel_names, scale_dtype, string_channel_names