+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ### Fix
 - Fixed aggregation issue when gene names are `NaN` or `None` (#101)
+- Fix Xenium reader for old Xenium data format (#105)
 
 ### Added
 - Support multipolygons in ROI rasterization
+- Added bins aggregation
+- Added Visium HD reader (tutorial comming soon)
 
 ## [1.1.2] - 2024-07-24